feat: move the EDU verify callout below the cycle toggle and extend EDU display to team stops
The verify nudge becomes a shared EduVerifyCallout card placed under the
billing-cycle toggle in both pricing tables. Team stops display the
stacked EDU total via a CreditSlider extra-discount prop, matching the
backend coupon math (5% of list on top of the volume discount). All
locales carry the EDU strings.
